extract `ReferencePoint` and transform nodes from `transform`
