+++ node-graph/README.md
@@ -157,7 +157,7 @@ raster_node!(graphene_core::raster::OpacityNode<_>, params: [f64]),
 
 There is also the more general `register_node!` for nodes that do not need to run per pixel.
 ```rs
-register_node!(graphene_core::transform::SetTransformNode<_>, input: VectorData, params: [DAffine2]),
+register_node!(graphene_core::transform_nodes::SetTransformNode<_>, input: VectorData, params: [DAffine2]),
 ```
 
 ## Debugging
