+++ node-graph/gcore/src/vector/mod.rs
@@ -2,11 +2,13 @@ mod algorithms;
 pub mod brush_stroke;
 pub mod generator_nodes;
 pub mod misc;
+mod reference_point;
 pub mod style;
 mod vector_data;
 mod vector_nodes;
 
 pub use bezier_rs;
+pub use reference_point::*;
 pub use style::PathStyle;
 pub use vector_data::*;
 pub use vector_nodes::*;
